fix(upload): selection not working immediately after adding keyword field
When a keyword field was added to the list of upload fields the selection of a keyword only started worked after reloading the whole app.

Re-subscribe to keyword selection events on each visit of the upload page to ensure the event callback is aware of changes in the form fields.

Closes #570