docs: add script and template for easier start of publication
To have less work at the end of the project, a script has been created, which copy all relevant documentation into one file. There are still manual steps to do, which are showed in the console.

Also a template for the word file has been created to have same formation over the whole file. And a macro is responsible formatting word tables

Closes #427